# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Urban Sanctuary API
  description: |
    Premium Apartment Booking Platform — Douala, Cameroon.

    The Urban Sanctuary API powers a modern apartment booking system with:
    - Multi-method authentication (Email/Password, Phone/OTP, 2FA)
    - Property listing management with image uploads
    - Booking engine with conflict prevention & loyalty points
    - Payment integration (MTN Mobile Money & Orange Money)
    - Role-based access control (Guest, Homeowner, Admin)
    - Review & rating system (verified stays only)
    - Full admin console with audit trail

    ## Authentication
    All authenticated endpoints require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    Access tokens are obtained via `/auth/login` or `/auth/register` and are valid for **15 minutes**.
    Refresh tokens are stored as HttpOnly cookies and are valid for **7 days**.

    ## Response Envelope
    All responses follow a standard JSON envelope:
    ```json
    {
      "success": true,
      "status": 200,
      "message": "...",
      "data": { ... },
      "meta": { "page": 1, "per_page": 20, "total": 100, "total_pages": 5 },
      "request_id": "uuid"
    }
    ```

    ## Rate Limiting
    - Global: 100 requests/minute per IP
    - Auth endpoints: 5 requests/minute per IP
    - Rate limit headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`
  version: 1.0.0
  contact:
...
  license:
...
servers:
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [Users]
      summary: Update current user profile
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /users/me/password:
    patch:
      tags: [Users]
      summary: Change password
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /users/me/loyalty:
    get:
      tags: [Users]
      summary: Get loyalty points balance
      description: |
        Returns the user's loyalty points balance and XAF equivalent.
        Each loyalty point = 10 XAF.
      operationId: getLoyalty
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Loyalty info retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoyaltyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/me/notifications:
    get:
      tags: [Users]
      summary: Get notifications
      operationId: getNotifications
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Notifications retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/NotificationResponse"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ──────────────────────────────────────────────
  # Properties
  # ──────────────────────────────────────────────
...
  /bookings/{id}/review:
    post:
      tags: [Bookings]
      summary: Submit a review
      description: |
        Submits a review for a completed booking. Only the guest can review.
        Reviews are automatically marked as verified since they are linked to a real stay.
        Only one review per booking is allowed.
      operationId: submitReview
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ReviewResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"

  # ──────────────────────────────────────────────
  # Payments
  # ──────────────────────────────────────────────
...
  /webhooks/mtn:
    post:
      tags: [Webhooks]
      summary: MTN MoMo payment callback
      description: |
        Receives payment status updates from MTN MoMo.
        Validates the `X-Callback-Signature` header.
        On successful payment: booking status → `confirmed`, 100 loyalty points awarded.
      operationId: mtnWebhook
      parameters:
        - name: X-Callback-Signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPayload"
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: processed
  /webhooks/orange:
    post:
      tags: [Webhooks]
      summary: Orange Money payment callback
      description: |
        Receives payment status updates from Orange Money.
        Validates the `X-Callback-Signature` header.
      operationId: orangeWebhook
      parameters:
        - name: X-Callback-Signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPayload"
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: processed

  # ──────────────────────────────────────────────
  # Admin
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from `/auth/login` or `/auth/register`
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        status:
          type: integer
          example: 200
        message:
          type: string
          example: Operation successful
        data:
          type: object
          nullable: true
        meta:
          $ref: "#/components/schemas/PaginationMeta"
        request_id:
          type: string
          format: uuid
      required: [success, status, message, request_id]
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        status:
          type: integer
          example: 400
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: One or more fields are invalid
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: email
                  message:
                    type: string
                    example: must be a valid email address
        request_id:
          type: string
          format: uuid
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          format: int64
          example: 150
        total_pages:
          type: integer
          example: 8

    # ── Auth ───────────────────────────────
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        avatar_url:
          type: string
          format: uri
          nullable: true
        role:
          type: string
          enum: [guest, homeowner, admin]
        is_verified:
          type: boolean
        totp_enabled:
          type: boolean
        loyalty_points:
          type: integer
          example: 500
        created_at:
          type: string
          format: date-time
    UpdateProfileRequest:
      type: object
      properties:
        first_name:
          type: string
          minLength: 1
          maxLength: 100
        last_name:
          type: string
          minLength: 1
          maxLength: 100
        avatar_url:
          type: string
          format: uri
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
        new_password:
          type: string
          minLength: 8
          maxLength: 128
        totp_code:
          type: string
          description: Required if 2FA is enabled
          minLength: 6
          maxLength: 6
    LoyaltyResponse:
      type: object
      properties:
        balance:
          type: integer
          example: 500
        points_value_xaf:
          type: number
          format: double
          description: XAF equivalent (1 point = 10 XAF)
          example: 5000.0
        history:
          type: array
          items:
            type: object
            properties:
              booking_id:
                type: string
                format: uuid
              type:
                type: string
                enum: [earned, redeemed]
              points:
                type: integer
              date:
                type: string
                format: date-time
    NotificationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          example: booking_confirmed
        channel:
          type: string
          enum: [email, sms, push]
        title:
          type: string
        content:
          type: string
        sent_at:
          type: string
          format: date-time
          nullable: true
        read_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    # ── Properties ─────────────────────────
...
    CreateReviewRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          maxLength: 2000
          example: Amazing apartment, great location!
    ReviewResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        booking_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        guest_id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        is_verified:
          type: boolean
          description: Always true (linked to a real stay)
        guest:
          $ref: "#/components/schemas/UserResponse"
        created_at:
          type: string
          format: date-time

    # ── Payments ───────────────────────────
...
    WebhookPayload:
      type: object
      description: Raw webhook payload from payment provider (structure varies)
      properties:
        externalId:
          type: string
          description: Our payment ID
        status:
          type: string
          description: "SUCCESSFUL, FAILED, or PENDING"
        financialTransactionId:
          type: string
          description: Provider's transaction reference

    # ── Admin ──────────────────────────────
...